refactor: split bff resource handlers
